Split workflow utils a bit
